refact  : plan list popup style
