+++ web/sections/pricing/pricing-table/index.tsx
@@ -34,90 +34,7 @@ const TeamPlanList = [
   "Unlimited Code blocks",
 ];
 
-function replaceStylePxToNumber(stylePx: string) {
-  return parseInt(stylePx.replace("px", ""));
-}
-
 export default function PlanList() {
-  const { addPopup, removePopup } = usePopupContext();
-  const width = useWindowWidth();
-
-  const handleClickQuestionMark = useCallback(() => {
-    addPopup({
-      title: "",
-      element: (
-        <FreePlanPopup width="100%" alignItems="center" px="25px" pb="60px">
-          {width >= replaceStylePxToNumber(defaultTheme.breakpoints[0]) && (
-            <Icon
-              name="faqClose"
-              ml="auto"
-              onClick={() => removePopup()}
-              style={{ cursor: "pointer" }}
-              mb="30px"
-              mt="20px"
-            />
-          )}
-
-          <Flex width="100%" alignItems="center" flexDirection="column">
-            <LandingpageText textAlign="center" variant="h4">
-              What are the limitations of free plan?
-            </LandingpageText>
-            <PopupDescription textAlign="center" variant="body1">
-              To build an enterprise level application, you’ll need a paid plan.
-              Paid plan includes extra default storage and unlimited projects
-              count. Also cloud objects such as translation token can be stored
-              up to 1 million. The extra usage will be charged as Standard Cloud
-              Fee.
-            </PopupDescription>
-          </Flex>
-        </FreePlanPopup>
-      ),
-    });
-  }, []);
-
-  const handleClickPaidPlan = useCallback(() => {
-    addPopup({
-      title: "",
-      element: (
-        <FreePlanPopup
-          width="100%"
-          alignItems="center"
-          px="24px"
-          pt="24px"
-          pb="48px"
-        >
-          {width >= replaceStylePxToNumber(defaultTheme.breakpoints[0]) && (
-            <Icon
-              name="faqClose"
-              ml="auto"
-              onClick={() => removePopup()}
-              style={{ cursor: "pointer" }}
-              mb="30px"
-            />
-          )}
-
-          <Flex width="100%" alignItems="center" flexDirection="column">
-            <LandingpageText textAlign="center" variant="h4">
-              Woopsy.
-            </LandingpageText>
-            <PopupDescription textAlign="center" variant="body1">
-              Bridged paid plans are disabled temporarily. Meanwhile, you can
-              use our free plan which basically does the same.
-              <Link href={LandingpageUrls.signup}>
-                <span
-                  className="cursor"
-                  style={{ color: "#172AD7", margin: "0px 5px" }}
-                >
-                  Sign up
-                </span>
-              </Link>
-              here.
-            </PopupDescription>
-          </Flex>
-        </FreePlanPopup>
-      ),
-    });
-  }, []);
 
   return (
     <SectionLayout alignContent="center">
